bug fixed, protected routes added

+++ src/contacts/contacts.controller.ts
@@ -1,6 +1,7 @@
-import { Body, Controller, Delete, Get, Param, ParseUUIDPipe, Post, Put } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Param, Post, Put, UnauthorizedException } from '@nestjs/common';
 import { ContactsService } from './contacts.service';
 import { CreateContactDto, GetContactsResponseDto, UpdateContactsDto } from './dto';
+import { User } from 'src/users/decorators/users.decorators';
 
 @Controller('contacts')
 export class ContactsController {
@@ -8,8 +9,9 @@ export class ContactsController {
     constructor(private readonly contactsService: ContactsService) {}
 
     @Get()
-    getAllContacts() {
-        return this.contactsService.getAllContacts();
+    getAllContacts(@User() userid: string) {
+        console.log(userid)
+        return this.contactsService.getAllContacts(userid);
     }
 
     @Get('/:id')
@@ -18,19 +20,25 @@ export class ContactsController {
         }
 
     @Post()
-    createContact(@Body() body: CreateContactDto) {
-        return this.contactsService.createContact(body);
+    createContact(@Body() body: CreateContactDto, @User() userid: string) {
+        return this.contactsService.createContact(body, userid);
     }
 
+
     @Put('/:id')
-    updateContact(
+    async updateContact(
         @Param('id') id: string,
-        @Body() body: UpdateContactsDto) {
+        @Body() body: UpdateContactsDto,
+        @User() userid: string) {
+            const contactUserId = await this.contactsService.getUserId(id);
+            if(contactUserId !== userid) throw new UnauthorizedException('You are not authorized to update this contact')
             return this.contactsService.updateContact(id, body);
         }
 
     @Delete('/:id')
-    deleteContact(@Param('id') id: string) {
+    async deleteContact(@Param('id') id: string, @User() userid: string) {
+        const contactUserId = await this.contactsService.getUserId(id);
+        if(contactUserId !== userid) throw new UnauthorizedException('You are not authorized to update this contact')
         return this.contactsService.deleteContact(id);
     }
